Extract all the TaskList and Storage into their own TaskList.java and Storage.java files
