--- codsworth.txt
@@ -1,1 +1,0 @@
-false/spacer/event/spacer/dwadas /from 2001-01-01T10:00 /to 2002-02-02T20:00
