+++ text-ui-test/EXPECTED.TXT
@@ -1,38 +1,53 @@
-    ____________________________________________________________
-    Hello, I'm Codsworth
-    What can I do for you?
-    ____________________________________________________________
-    ____________________________________________________________
-    Got it. I've added this task:
-      [T][ ] borrow book
-    Now you have 1 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-    Here are the tasks in your list:
-    1.[T][ ] borrow book
-    ____________________________________________________________
-    ____________________________________________________________
-    Got it. I've added this task:
-      [D][ ] return book (by: 20 Jul 2001 01:00pm)
-    Now you have 2 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-    Got it. I've added this task:
-      [E][ ] project meeting (from: 21 Jul 2001 to: 22 Aug 2002 02:00pm)
-    Now you have 3 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-    Got it. I've added this task:
-      [D][ ] do homework (by: 10 Oct 1010)
-    Now you have 4 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-    Here are the tasks in your list:
-    1.[T][ ] borrow book
-    2.[D][ ] return book (by: 20 Jul 2001 01:00pm)
-    3.[E][ ] project meeting (from: 21 Jul 2001 to: 22 Aug 2002 02:00pm)
-    4.[D][ ] do homework (by: 10 Oct 1010)
-    ____________________________________________________________
-    ____________________________________________________________
-    Bye. Hope to see you again soon!
-    ____________________________________________________________
+________________________________________________________
+|   _____          _                        _   _      |
+|  / ____|        | |                      | | | |     |
+| | |     ___   __| |_____      _____  _ __| |_| |__   |
+| | |    / _ \ / _` / __\ \ /\ / / _ \| '__| __| '_ \  |
+| | |___| (_) | (_| \__ \\ V  V / (_) | |  | |_| | | | |
+|  \_____\___/ \__,_|___/ \_/\_/ \___/|_|   \__|_| |_| |
+________________________________________________________
+Hello there, my name is Codsworth
+What can I do for you?
+________________________________________________________
+________________________________________________________
+Got it, I've added this task:
+[T][ ] borrow book
+You have 1 task(s) left.
+________________________________________________________
+________________________________________________________
+Here are all the task(s) in your list:
+1. [T][ ] borrow book
+________________________________________________________
+________________________________________________________
+Got it, I've added this task:
+[D][ ] return book (by: 20 Jul 2001 01:00pm)
+You have 2 task(s) left.
+________________________________________________________
+________________________________________________________
+Got it, I've added this task:
+[E][ ] project meeting (from: 21 Jul 2001 to: 22 Aug 2002 02:00pm)
+You have 3 task(s) left.
+________________________________________________________
+________________________________________________________
+Got it, I've added this task:
+[D][ ] do homework (by: 10 Oct 1010)
+You have 4 task(s) left.
+________________________________________________________
+________________________________________________________
+Here are all the task(s) in your list:
+1. [T][ ] borrow book
+2. [D][ ] return book (by: 20 Jul 2001 01:00pm)
+3. [E][ ] project meeting (from: 21 Jul 2001 to: 22 Aug 2002 02:00pm)
+4. [D][ ] do homework (by: 10 Oct 1010)
+________________________________________________________
+________________________________________________________
+|   _____          _                        _   _      |
+|  / ____|        | |                      | | | |     |
+| | |     ___   __| |_____      _____  _ __| |_| |__   |
+| | |    / _ \ / _` / __\ \ /\ / / _ \| '__| __| '_ \  |
+| | |___| (_) | (_| \__ \\ V  V / (_) | |  | |_| | | | |
+|  \_____\___/ \__,_|___/ \_/\_/ \___/|_|   \__|_| |_| |
+________________________________________________________
+Thank you for using Codsworth
+Hope to see you soon!
+________________________________________________________
